enable the post-upgrade "Optimizing apps" progress indication UI
Adjustment to ViewRootImpl is needed because AudioService is not yet started when this UI is shown,
which leads to a system_server crash.

Caching the result of AudioManager.areNavigationRepeatSoundEffectsEnabled() is not necessary, this
method is fast and its result is used only for sound effects when using keyboard for navigation.

Signed-off-by: Dmitrii <bankersenator@gmail.com>
(cherry picked from commit babe037fd36467cf87795c3eacb6abfa39664c7e)